refactor(doc): generalize `post_process_templ_links()`
